Client Requests

EasyCurl.curl_open — Function

curl_open(f::Function, x...; kw...)

A helper function for executing a batch of curl requests, using the same client. Optionally configure the client (see CurlClient for more details).

Examples

julia> curl_open() do client
           response = curl_request(client, "GET", "http://httpbin.org/get")
           println(curl_status(response))
       end
200

EasyCurl.curl_request — Function

curl_request(method::AbstractString, url::AbstractString; kw...) -> CurlResponse

Send a url HTTP CurlRequest using as method one of "GET", "POST", etc. and return a CurlResponse object.

Keyword arguments

headers = Pair{String,String}[]: The headers for the request.

body = nothing: The body for the request.

query = nothing: The query string for the request.

interface = nothing: The interface for the request.

status_exception = true: Whether to throw an exception if the response status code indicates an error.

connect_timeout = 60: The connect timeout for the request in seconds.

read_timeout = 300: The read timeout for the request in seconds.

retry = 1: The number of times to retry the request if an error occurs.

proxy = nothing: Which proxy to use for the request.

accept_encoding = "gzip": Encoding to accept.

verbose::Bool = false: Enables verbose output from Curl for debugging.

ssl_verifypeer = true: Whether peer need to be verified.

Examples

julia> headers = Pair{String,String}[
           "User-Agent" => "EasyCurl.jl",
           "Content-Type" => "application/json"
       ]

julia> response = curl_request("POST", "http://httpbin.org/post", headers = headers, query = "qry=你好嗎",
    body = "{\"data\":\"hi\"}", interface = "en0", read_timeout = 5, connect_timeout = 10, retry = 10)

julia> curl_status(response)
200

julia> curl_body(response) |> String |> print
{
  "headers": {
    "X-Amzn-Trace-Id": "Root=1-6588a009-19f3dc0321bee38106226bb3",
    "Content-Length": "13",
    "Host": "httpbin.org",
    "Accept": "*/*",
    "Content-Type": "application/json",
    "Accept-Encoding": "gzip",
    "User-Agent": "EasyCurl.jl"
  },
  "json": {
    "data": "hi"
  },
  "files": {},
  "args": {
    "qry": "你好嗎"
  },
  "data": "{\"data\":\"hi\"}",
  "url": "http://httpbin.org/post?qry=你好嗎",
  "form": {},
  "origin": "100.250.50.140"
}

EasyCurl.curl_get — Function

curl_get(url::AbstractString; kw...) -> CurlResponse

Shortcut for curl_request function, work similar to curl_request("GET", url; kw...).

Examples

julia> headers = Pair{String,String}[
           "User-Agent" => "EasyCurl.jl",
           "Content-Type" => "application/json"
       ]

julia> response = curl_get("http://httpbin.org/get", headers = headers,
    query = Dict{String,String}("qry" => "你好嗎"))

julia> curl_status(response)
200

julia> curl_body(response) |> String |> print
{
  "args": {
    "qry": "你好嗎"
  },
  "headers": {
    "Accept": "*/*",
    "Accept-Encoding": "gzip",
    "Content-Type": "application/json",
    "Host": "httpbin.org",
    "User-Agent": "EasyCurl.jl",
    "X-Amzn-Trace-Id": "Root=1-6589e259-24815d6d62da962a06fc7edf"
  },
  "origin": "100.250.50.140",
  "url": "http://httpbin.org/get?qry=你好嗎"
}

EasyCurl.curl_patch — Function

curl_patch(url::AbstractString; kw...) -> CurlResponse

Shortcut for curl_request function, work similar to curl_request("PATCH", url; kw...).

Examples

julia> headers = Pair{String,String}[
           "User-Agent" => "EasyCurl.jl",
           "Content-Type" => "application/json"
       ]

julia> response = curl_patch("http://httpbin.org/patch", headers = headers,
    query = "qry=你好嗎", body = "{\"data\":\"hi\"}")

julia> curl_status(response)
200

julia> curl_body(response) |> String |> print
{
  "args": {
    "qry": "你好嗎"
  },
  "data": "{\"data\":\"hi\"}",
  "files": {},
  "form": {},
  "headers": {
    "Accept": "*/*",
    "Accept-Encoding": "gzip",
    "Content-Length": "13",
    "Content-Type": "application/json",
    "Host": "httpbin.org",
    "User-Agent": "EasyCurl.jl",
    "X-Amzn-Trace-Id": "Root=1-6589e410-33f8cb5a31db9fba6c0a746f"
  },
  "json": {
    "data": "hi"
  },
  "origin": "100.250.50.140",
  "url": "http://httpbin.org/patch?qry=你好嗎"
}

EasyCurl.curl_post — Function

curl_post(url::AbstractString; kw...) -> CurlResponse

Shortcut for curl_request function, work similar to curl_request("POST", url; kw...).

Examples

julia> headers = Pair{String,String}[
           "User-Agent" => "EasyCurl.jl",
           "Content-Type" => "application/json"
       ]

julia> response = curl_post("http://httpbin.org/post", headers = headers,
    query = "qry=你好嗎", body = "{\"data\":\"hi\"}")

julia> curl_status(response)
200

julia> curl_body(response) |> String |> print
{
  "args": {
    "qry": "你好嗎"
  },
  "data": "{\"data\":\"hi\"}",
  "files": {},
  "form": {},
  "headers": {
    "Accept": "*/*",
    "Accept-Encoding": "gzip",
    "Content-Length": "13",
    "Content-Type": "application/json",
    "Host": "httpbin.org",
    "User-Agent": "EasyCurl.jl",
    "X-Amzn-Trace-Id": "Root=1-6589e32c-7f09b85d56e11aea59cde1d6"
  },
  "json": {
    "data": "hi"
  },
  "origin": "100.250.50.140",
  "url": "http://httpbin.org/post?qry=你好嗎"
}

EasyCurl.curl_put — Function

curl_put(url::AbstractString; kw...) -> CurlResponse

Shortcut for curl_request function, work similar to curl_request("PUT", url; kw...).

Examples

julia> headers = Pair{String,String}[
           "User-Agent" => "EasyCurl.jl",
           "Content-Type" => "application/json"
       ]

julia> response = curl_put("http://httpbin.org/put", headers = headers,
    query = "qry=你好嗎", body = "{\"data\":\"hi\"}")

julia> curl_status(response)
200

julia> curl_body(response) |> String |> print
{
  "args": {
    "qry": "你好嗎"
  },
  "data": "{\"data\":\"hi\"}",
  "files": {},
  "form": {},
  "headers": {
    "Accept": "*/*",
    "Accept-Encoding": "gzip",
    "Content-Length": "13",
    "Content-Type": "application/json",
    "Host": "httpbin.org",
    "User-Agent": "EasyCurl.jl",
    "X-Amzn-Trace-Id": "Root=1-6589e3b0-58cdde84399ad8be30eb4e46"
  },
  "json": {
    "data": "hi"
  },
  "origin": "100.250.50.140",
  "url": "http://httpbin.org/put?qry=你好嗎"
}

EasyCurl.curl_head — Function

curl_head(url::AbstractString; kw...) -> CurlResponse

Shortcut for curl_request function, work similar to curl_request("HEAD", url; kw...).

Examples

julia> headers = Pair{String,String}[
           "User-Agent" => "EasyCurl.jl",
           "Content-Type" => "application/json"
       ]

julia> response = curl_head("http://httpbin.org/get", headers = headers,
    query = "qry=你好嗎", interface = "0.0.0.0")

julia> curl_status(response)
200

julia> curl_body(response)
UInt8[]

EasyCurl.curl_delete — Function

curl_delete(url::AbstractString; kw...) -> CurlResponse

Shortcut for curl_request function, work similar to curl_request("DELETE", url; kw...).

Examples

julia> headers = Pair{String,String}[
           "User-Agent" => "EasyCurl.jl",
           "Content-Type" => "application/json"
       ]

julia> response = curl_delete("http://httpbin.org/delete", headers = headers,
    query = "qry=你好嗎", body = "{\"data\":\"hi\"}")

julia> curl_status(response)
200

julia> curl_body(response) |> String |> print
{
  "args": {
    "qry": "你好嗎"
  },
  "data": "{\"data\":\"hi\"}",
  "files": {},
  "form": {},
  "headers": {
    "Accept": "*/*",
    "Accept-Encoding": "gzip",
    "Content-Length": "13",
    "Content-Type": "application/json",
    "Host": "httpbin.org",
    "User-Agent": "EasyCurl.jl",
    "X-Amzn-Trace-Id": "Root=1-6589e5f7-1c1ff2407f567ff17786576d"
  },
  "json": {
    "data": "hi"
  },
  "origin": "100.250.50.140",
  "url": "http://httpbin.org/delete?qry=你好嗎"
}

Client/Request/Response types

EasyCurl.CurlClient — Type

CurlClient

Represents a client for making HTTP requests using libcurl. Allows for connection reuse.

Fields

curl_handle::Ptr{CURL}: The libcurl easy handle.

multi_handle::Ptr{CURL}: The libcurl multi handle.

EasyCurl.CurlRequest — Type

CurlRequest

Represents an HTTP request object.

Fields

method::String: The HTTP method for the request (e.g., "GET", "POST").

url::String: The target URL for the HTTP request.

method::String: Specifies the HTTP method for the request (e.g., "GET", "POST").

url::String: The target URL for the HTTP request.

headers::Vector{Pair{String, String}}: A list of header key-value pairs to include in the request.

body::Vector{UInt8}: The body of the request, represented as a vector of bytes.

connect_timeout::Real: Timeout in seconds for establishing a connection.

read_timeout::Real: Timeout in seconds for reading the response.

interface::Union{String,Nothing}: Specifies a particular network interface to use for the request, or nothing to use the default.

proxy::Union{String,Nothing}: Specifies a proxy server to use for the request, or nothing to bypass proxy settings.

accept_encoding::Union{String,Nothing}: Specifies the accepted encodings for the response, such as "gzip". Defaults to nothing if not set.

ssl_verifypeer::Bool: Indicates whether SSL certificates should be verified (true) or not (false).

verbose::Bool: When true, enables detailed output from Curl, useful for debugging purposes.

EasyCurl.CurlContext — Type

CurlContext

A service object for managing the state of a Curl request.

Fields

curl_slist_ptr::Ptr{Nothing}: A pointer to a linked list of headers.

curl_status::Vector{Clong}: The HTTP status code of the response.

curl_total_time::Vector{Cdouble}: The time taken for the HTTP request in seconds.

curl_active::Vector{Cint}: The number of active requests.

byte_received::UInt64: The number of bytes received.

headers::IOBuffer: The headers received in the HTTP response.

body::IOBuffer: The body of the response.

EasyCurl.CurlResponse — Type

CurlResponse(x::CurlContext)

An HTTP response object returned on a request completion. Constructed from a CurlContext object.

Fields

status::Int64: The HTTP status code of the response.

request_time::Float64: The time taken for the HTTP request in seconds.

headers::Vector{Pair{String,String}}: Headers received in the HTTP response.

body::Vector{UInt8}: The response body as a vector of bytes.

See also: curl_status, curl_request_time, curl_headers, curl_body, curl_iserror

Utilities

EasyCurl.urlencode — Function

EasyCurl.urlencode(s::AbstractString)

Encode a string s into URI using only the US-ASCII characters legal within a URI.

Examples

julia> EasyCurl.urlencode("[curl]")
"%5Bcurl%5D"

EasyCurl.urldecode — Function

EasyCurl.urldecode(s::AbstractString)

Decode an encoded URI string s back to normal string.

Examples

julia> EasyCurl.urldecode("%5Bcurl%5D")
"[curl]"

EasyCurl.joinurl — Function

EasyCurl.joinurl(basepart::AbstractString, parts::AbstractString...)::String

Construct a URL by concatenating a base part with one or more path segments. This function ensures that each segment is separated by a single forward slash (/), regardless of whether the basepart or parts already contain slashes at their boundaries.

Examples

julia> EasyCurl.joinurl("http://example.com", "path")
"http://example.com/path"

julia> EasyCurl.joinurl("http://example.com/", "/path/to/resource")
"http://example.com/path/to/resource"

Advanced Topics

EasyCurl.curl_body — Function

curl_body(x::CurlResponse) -> Vector{UInt8}

Returns the response body from a CurlResponse object.

EasyCurl.curl_status — Function

curl_status(x::CurlResponse) -> Int64

Returns the HTTP status code from a CurlResponse object.

EasyCurl.curl_headers — Function

curl_headers(x::CurlResponse) -> Vector{Pair{String,String}}

Returns the HTTP headers from a CurlResponse object.

EasyCurl.curl_request_time — Function

curl_request_time(x::CurlResponse) -> Float64

Returns the request time from a CurlResponse object.

EasyCurl.curl_iserror — Function

curl_iserror(x::CurlResponse) -> Bool

Checks whether CurlResponse contains an errorneous HTTP status code (>= 400).