Curl Interface

Types

EasyCurl.CurlResponse — Type

CurlResponse

Common response interface for all supported protocols.

Interface accessors

curl_total_time(x): Total time spent receiving a response in seconds.

curl_body(x): The response body.

EasyCurl.CurlClient — Type

CurlClient

Represents a client for making HTTP requests using libcurl. Allows for connection reuse.

Fields

curl_handle::Ptr{Cvoid}: The libcurl easy handle.

Methods

Base.close — Function

close(client::CurlClient)

Closes the client instance by cleaning up the associated libcurl easy handle.

EasyCurl.curl_open — Function

curl_open(f::Function, x...; kw...)

A helper function for executing a batch of curl requests, using the same client. Optionally configure the client (see CurlClient for more details).

Examples

julia> curl_open() do client
           response = http_request(client, "GET", "http://httpbin.org/get")
           curl_status(response)
       end
200